src/blob-om-py/requirements.txt
azure-storage-blob
azure-identity
azure-mgmt-storage
azure-mgmt-resource